Updated RootScanner::UpdateZoom.
The UpdateZoom now only takes an argument for a TVirtualPad* and scans
the items in that pad for TGraph and TH1 objects. These are then used to
compute the pad limits. The zoom controls are handled in a similar way.

This updated caused a minor change in the call from scope.

This update permitted the autozooming the work in a fashion similar to
the expectation for HistScanner. In addition, it permitted a substantial
reduction int he code overhead in HistScanner.


Former-commit-id: abfea8a3a430a2beee85d307b197e70ca0dcb938