Changed plot expression validation.
The expr (previously type) is now validated using a TTreeFormula. This
will permit extension of the data available to the user without
requiring redundant list of valid parameters at the plot command. In
addition, it will open the door to more complicated plotting include 2D
histograms.

TTreeFormula is in the libTreePlayer library which is not linked by
default, this required a change in the CMake file.


Former-commit-id: f47f40407a3a974af8491ad032b942ba3141ff1b